Implement naive gene UUIDs
These UUIDs are only used when
clojush.random/random-plush-instruction-map is called. Alternation will
work as expected, but mutation and {close,silent} mutation need to be
looked at. At this point, we need to configure logging and then we can
do some alternation only runs and experiement with loading them into
a database.