Update handling of too-big-child
This commit should not change the evolutionary setup in any way, but it
is a non-trivial change to logging. Previously, if a child's geneome was
too long one of four things happened.

1. The child became an exact copy of the parent. Not just the genome,
   but *everything*.
2. The child became empty, with :genetic-operators set to :empty.
3. The child's genome was truncated and :genetic-operators was left
   empty.
4. The child was replaced with a random one and :genetic-operators was
   set to :random.

I think I got that right, but the point is that it was rather
inconsistent. Now, regardless of which of the four is chosen,
:genetic-operators is set to the operator that ostensibly created it. If
the child is replaced with a random one, :is-random-replacement is set
to true. This mirrors the behavior of the autoconstruction operator.